msgpack-bind/src/test/java/com/lisb/msgpack/bind/marshaller/LongMarshallerTest.java
package com.lisb.msgpack.bind.marshaller;

import com.lisb.msgpack.bind.Marshaller;
import com.lisb.msgpack.bind.Unmarshaller;
import com.lisb.msgpack.bind.unmarshaller.BooleanUnmarshaller;
import com.lisb.msgpack.bind.unmarshaller.LongUnmarshaller;
import org.junit.Assert;
import org.junit.BeforeClass;
import org.junit.Test;
import org.msgpack.core.MessageBufferPacker;
import org.msgpack.core.MessagePack;

public class LongMarshallerTest {
    private static Marshaller<Long> marshaller;
    private static Unmarshaller<Long> unmarshaller;

    @BeforeClass
    public static void beforeClass() {
        marshaller = LongMarshaller.getInstance();
        unmarshaller = LongUnmarshaller.getInstance();
    }

    @Test
    public void testMarshalAndUnmarshal() throws Exception {
        final MessageBufferPacker packer = MessagePack.newDefaultBufferPacker();

        final Long source = 55454361645L;

        marshaller.marshal(packer, source);

        final Long dest = unmarshaller.unmarshal(MessagePack.newDefaultUnpacker(packer.toByteArray()));

        Assert.assertEquals(source, dest);
    }

    @Test
    public void testNullBMarshalAndUnmarshal() throws Exception {
        final MessageBufferPacker packer = MessagePack.newDefaultBufferPacker();
        marshaller.marshal(packer, null);
        Assert.assertNull(unmarshaller.unmarshal(MessagePack.newDefaultUnpacker(packer.toByteArray())));
    }
}
